% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategorizeReads.R
\name{CategorizeReads}
\alias{CategorizeReads}
\title{Categorizes the reads in one sam file.}
\usage{
CategorizeReads(sam, POS, REF, ALT)
}
\arguments{
\item{sam}{An in-memory represenation of a sam file as data.frame.}

\item{POS}{The position in the sam at which the DBS starts.}

\item{REF}{The reference allele.}

\item{ALT}{The putative alternative allele.}
}
\value{
a named character vector, the names are the read names in the sam file.
Each element is one of
"Read supports only 1st position"
"Read supports only 2nd position",
"WT read"
"Mut read"
}
\description{
Categorizes the reads in one sam file.
}
