% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBAMSlice.R
\name{CreateBAMSlice}
\alias{CreateBAMSlice}
\title{Get a BAM slice as a \code{data.frame} possibly using different functions.}
\usage{
CreateBAMSlice(
  BAM.name,
  CHROM,
  POS,
  padding = 10,
  read.fn = CreateBAMSliceFileSamtools
)
}
\arguments{
\item{BAM.name}{The input BAM file name.}

\item{CHROM}{The chromosome from which to slice.}

\item{POS}{The position in \code{CHROM} from which to slice.}

\item{padding}{The number of basepairs on each side of \code{POS} to take.}

\item{read.fn}{The function to use to get the BAM slice as a SAM file.}
}
\value{
The SAM file contents as a \code{data.frame}
}
\description{
Get a BAM slice as a \code{data.frame} possibly using different functions.
}
