% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBAMSliceFileSamtools.R
\name{CreateBAMSliceFileSamtools}
\alias{CreateBAMSliceFileSamtools}
\title{Create a BAM slice as a SAM file using the samtools executable.}
\usage{
CreateBAMSliceFileSamtools(
  BAM.name,
  CHROM,
  POS,
  padding = 10,
  save.file.path = NULL
)
}
\arguments{
\item{BAM.name}{The input BAM file name.}

\item{CHROM}{The chromosome from which to slice.}

\item{POS}{The position in \code{CHROM} from which to slice.}

\item{padding}{The number of basepairs on each side of \code{POS} to take.}

\item{save.file.path}{If not NULL, save the file to that file path.}
}
\value{
The path of the SAM file that contains the BAM slice.
}
\description{
Create a BAM slice as a SAM file using the samtools executable.
}
