% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBS_conclusion_1_row.R
\name{DBS_conclusion_1_row}
\alias{DBS_conclusion_1_row}
\title{Examine 1 row of DBS VCF already processed by \code{\link{Slice2ReadSupport}} and decide which DBSs are real.}
\usage{
DBS_conclusion_1_row(row, germlineCutOff = 0.2, max.half.support.T.reads = 1)
}
\arguments{
\item{row}{One row of a DBS VCF (as a \code{data.frame})
already processed by \code{\link{Slice2ReadSupport}}  (so that the fields
\code{vcf$NreadSupport} and \code{vcf$TreadSupport} are populated).}

\item{germlineCutOff}{If this proportion of normal reads show one or the
other variant (or both variants), consider this a germline variant or
partial germline variant.}

\item{max.half.support.T.reads}{Do not tolerate more than this number of reads in the tumor
that support one but not both mutated positions.}
}
\value{
A character string indicating the conclusion about the putative DBS.
}
\description{
Examine 1 row of DBS VCF already processed by \code{\link{Slice2ReadSupport}} and decide which DBSs are real.
}
