% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllBAMSlicesSamtools.R
\name{GetAllBAMSlicesSamtools}
\alias{GetAllBAMSlicesSamtools}
\title{Get and save all slices of BAM as specified by positions in a VCF table.}
\usage{
GetAllBAMSlicesSamtools(
  vcf,
  bam.name,
  padding = 10,
  where.to.put.slices = tempfile(),
  verbose = 0
)
}
\arguments{
\item{vcf}{A VCF (Variant Call Format) "file" as a data.table or similar.}

\item{bam.name}{The file name (path) to the BAM file to slice.}

\item{padding}{How many base pairs on each side of the first base of the DBS
to to keep in the BAM slices.}

\item{where.to.put.slices}{If \code{NULL}, create a temporary directory to store
the slices.  Otherwise, a character string that specifies a directory in
which to store the BAM slices. This is directory is created if necessary.}

\item{verbose}{If > 0 print a message when starting the number of slices
generated every \code{verbose} slices.}
}
\value{
A character string that specifies the directory the containing the
BAM slices. The slices are stored as SAM files.
}
\description{
Get and save all slices of BAM as specified by positions in a VCF table.
}
