% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReadSupport.R
\name{GetReadSupport}
\alias{GetReadSupport}
\title{Calculate the read support for putative DBS a given position in a BAM file}
\usage{
GetReadSupport(BAM.name, CHROM, POS, read.fn = CreateBAMSliceFileSamtools)
}
\arguments{
\item{BAM.name}{The input BAM file name.}

\item{CHROM}{The chromosome from which to slice.}

\item{POS}{The position in \code{CHROM} from which to slice.}

\item{read.fn}{The function to use to get the BAM slice as a SAM file.}
}
\description{
Calculate the read support for putative DBS a given position in a BAM file
}
