% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_DBS_VCF_and_BAMs_to_verfiy_DBSs.R
\name{Read_DBS_VCF_and_BAMs_to_verfiy_DBSs}
\alias{Read_DBS_VCF_and_BAMs_to_verfiy_DBSs}
\title{Determine whether sequencing reads in fact support DBSs present in a VCF file.}
\usage{
Read_DBS_VCF_and_BAMs_to_verfiy_DBSs(
  input.vcf,
  Nbam.name,
  Tbam.name,
  N.slice.dir = tempfile(),
  T.slice.dir = tempfile(),
  unlink.slice.dir = TRUE,
  exclude.SBSs = TRUE,
  verbose = 0,
  outfile = NULL
)
}
\arguments{
\item{input.vcf}{If a character string, then the path to a VCF file; otherwise
A a single VCF "file" as a data.frame or similar object.}

\item{Nbam.name}{The name of the BAM file for the normal sample corresponding to \code{vcf.name}.}

\item{Tbam.name}{The name of the BAM file for the tumor sample corresponding to \code{vcf.name}.}

\item{N.slice.dir}{Directory for the slices of the normal BAM.
Created if necessary.}

\item{T.slice.dir}{Directory for the slices of the tumor BAM.
Created if necessary. Must be different than \code{N.slice.dir}.}

\item{unlink.slice.dir}{If \code{TRUE} unlink \code{N.slice.dir}
and \code{T.slice.dir} before return.}

\item{exclude.SBSs}{If \code{TRUE} silently filter out (exclude)
SBSs in the input VCF. This makes sense if the the VCF is from
a caller (like Mutect or the Hartwig Medical Foundation caller)
that calls both SBSs and DBS.}

\item{verbose}{If > 0 print a message when starting the number of slices
generated every \code{verbose} slices.}

\item{outfile}{If not \code{NULL} then write the "evaluated" VCF to \code{outfile};
otherwise write it to \code{paste0(input.vcf(vcf.name, "_evaluated.vcf")}. Must be
non-\code{NULL} if \code{} is not a file path.}
}
\value{
Invisibly, a list with the elements
\enumerate{
\item The name of the DBS-only VCF file created.
\item The in-memory representation of the DBS VCF as a \code{data.table}.
\item The directory with the normal sam slices, if \code{unlink.slice.dir} is \code{FALSE}.
\item The directory with the tumor sam slices, if \code{unlink.slice.dir} is \code{FALSE}.
}
}
\description{
Determine whether sequencing reads in fact support DBSs present in a VCF file.
}
\details{
Creates a new VCF file.
This VCF file has no data rows if there were no DBSs to analyze.
Otherwise, this VCF contains the additional columns.
Any SBSs or indels in the input are silently ignored,
and no attempt is made to merge adjacent SBSs.
\enumerate{
\item \code{Format} contains the fixed string \code{"WtReads:pos1reads:pos2reads:MutReads"}.
\item \code{NreadSupport} With regard to the two positions of the DBS in
the normal BAM, a string with 4 numbers separated by ":", with the numbers
indicating respectively:
\itemize{
\item the number of reads that are reference sequence at
both positions of the DBS,
\item the number of reads that that have the alternative
allele only at the 1st position of the DBS,
\item the number of reads that
have the alternative allele only at the second position of the DBS, and
\item the number
of reads that have the alternative alleles at both positions of the DBS.
}
\item \code{TreadSupport} Information analogous to that in \code{NreadSupport}, for the
tumor BAM.
\item \code{DBSconclusion} A string that describes whether the DBSs is
believable (\code{"True DBS"}), and if not, a string that describes
why not.
}
}
