% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_SBS_VCF_and_BAMs_to_verify_DBSs.R
\name{Read_SBS_VCF_and_BAMs_to_verify_DBSs}
\alias{Read_SBS_VCF_and_BAMs_to_verify_DBSs}
\title{Aggressively merge SBSs into DBSs then determine whether sequencing reads in fact support DBSs.}
\usage{
Read_SBS_VCF_and_BAMs_to_verify_DBSs(
  input.vcf,
  Nbam.name,
  Tbam.name,
  variant.caller,
  N.slice.dir = tempfile(),
  T.slice.dir = tempfile(),
  unlink.slice.dir = TRUE,
  outfile = NULL,
  verbose = 0
)
}
\arguments{
\item{input.vcf}{If a character string, then the path to a VCF file; otherwise
A a single VCF "file" as a data.frame or similar object.}

\item{Nbam.name}{The name of the BAM file for the normal sample corresponding to \code{vcf.name}.}

\item{Tbam.name}{The name of the BAM file for the tumor sample corresponding to \code{vcf.name}.}

\item{variant.caller}{One of \code{"strelka"}, \code{"PCAWG"},
or \code{"unknown"}.
Merging adjacent SBS is done by \code{ReadAndSplitVCFs} in the ICAMS package.}

\item{N.slice.dir}{Directory for the slices of the normal BAM.
Created if necessary.}

\item{T.slice.dir}{Directory for the slices of the tumor BAM.
Created if necessary. Must be different than \code{N.slice.dir}.}

\item{unlink.slice.dir}{If \code{TRUE} unlink \code{N.slice.dir}
and \code{T.slice.dir} before return.}

\item{outfile}{If not \code{NULL} then write the "evaluated" VCF to \code{outfile};
otherwise write it to \code{paste0(input.vcf(vcf.name, "_evaluated.vcf")}. Must be
non-\code{NULL} if \code{input.vcf} is not a file path.}

\item{verbose}{If > 0 print a message when starting the number of slices
generated every \code{verbose} slices.}
}
\value{
Same as \code{\link{Read_DBS_VCF_and_BAMs_to_verify_DBSs}}.
}
\description{
Aggressively merge SBSs into DBSs then determine whether sequencing reads in fact support DBSs.
}
\details{
Note: argument \code{input.vcf} must be a file path.
This function creates a new VCF file.
See \code{\link{Read_DBS_VCF_and_BAMs_to_verify_DBSs}}
for details.
}
