% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVCFAndBAMsAndProcess.R
\name{ReadVCFAndBAMsAndProcess}
\alias{ReadVCFAndBAMsAndProcess}
\title{Determine whether sequencing reads in fact support DBSs generated by processing a VCF file.}
\usage{
ReadVCFAndBAMsAndProcess(
  vcf.name,
  Nbam.name,
  Tbam.name,
  variant.caller,
  num.cores = 10,
  outfile = NULL
)
}
\arguments{
\item{vcf.name}{The name of the VCF file to analyze.}

\item{Nbam.name}{The name of the BAM file for the normal sample corresponding to \code{vcf.name}.}

\item{Tbam.name}{The name of the BAM file for the tumor sample corresponding to \code{vcf.name}.}

\item{variant.caller}{One of \code{"strelka"}, \code{"mutect"}, \code{"PCAWG"}, or \code{"unknown"}.
If \code{variant.caller} is \code{"unknown"}, then the variant caller must have called
DBSs as such. For \code{"strelka"} and {"mutect"} DBS analysis is as in \code{ICAMS};
Strelka does not call DBSs, and DBSs are inferred from adjacent single base substitutions.
This is hte case for \code{"PCAWG"} as well.}

\item{num.cores}{Number of cores to use while reading the VCF. Set this to 1 for testing
and on MS Windows.}

\item{outfile}{If not \code{NULL} then write the "evaluated" VCF to \code{outfile};
otherwise write it to \code{paste0(vcf.name, "_evaluated.vcf")}.}
}
\value{
Invisibly, a list, the first element of which is the name of the file created, the
second of which is an in-memory representation of the VCF as a \code{data.table}.
}
\description{
Determine whether sequencing reads in fact support DBSs generated by processing a VCF file.
}
\details{
Creates a new VCF file.
The additional information in this VCF files is described in \code{\link{VerifyDBSVcf}}.
}
