% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVCFAndBAMsAndVerifyDBSs.R
\name{ReadVCFAndBAMsAndVerifyDBSs}
\alias{ReadVCFAndBAMsAndVerifyDBSs}
\title{Determine whether sequencing reads in fact support DBSs generated by processing a VCF file.}
\usage{
ReadVCFAndBAMsAndVerifyDBSs(
  input.vcf.name,
  Nbam.name,
  Tbam.name,
  variant.caller,
  always.merge.SBS = FALSE,
  num.cores = 10,
  N.slice.dir = tempfile(),
  T.slice.dir = tempfile(),
  unlink.slice.dir = TRUE,
  outfile = NULL,
  verbose = 0
)
}
\arguments{
\item{input.vcf.name}{The name of the VCF file to analyze. The VCF file can be gzip'ed.}

\item{Nbam.name}{The name of the BAM file for the normal sample corresponding to \code{vcf.name}.}

\item{Tbam.name}{The name of the BAM file for the tumor sample corresponding to \code{vcf.name}.}

\item{variant.caller}{One of \code{"strelka"}, \code{"mutect"}, \code{"PCAWG"},
"HMF", or \code{"unknown"}.
If \code{variant.caller} is \code{"unknown"}, then the variant caller must have called
DBSs as such. For \code{"strelka"} and {"mutect"}.
\code{"PCAWG"} indicates the format in the consensus VCFs from the PCAWG project.
\code{"HMF"} indicates the format used by the Hartwig Medical Foundation
variant caller.
For \code{"mutect"} and \code{"HMF"}, only the DBSs called by the variant caller
are analyzed. Adjacent SBS are never merged.
For \code{"strelka"} adjacent SBSs are always merged.
For \code{"unknown"} adjacent SBSs are merged if and only if
\code{always.merge.SBS = TRUE}.
DBS analysis is done by \code{\link[ICAMS]{ReadAndSplitVCFs}}.}

\item{always.merge.SBS}{Merge adjacent SBS into DBSs
if \code{variant.caller = "uknown"}. Otherwise silently ignored.}

\item{num.cores}{Number of cores to use while reading the VCF. Set this to 1 for testing
and on MS Windows.}

\item{N.slice.dir}{Directory for the slices of the normal BAM.
Created if necessary.}

\item{T.slice.dir}{Directory for the slices of the tumor BAM.
Created if necessary. Must be different than \code{N.slice.dir}.}

\item{unlink.slice.dir}{If \code{TRUE} unlink \code{N.slice.dir}
and \code{T.slice.dir} before return.}

\item{outfile}{If not \code{NULL} then write the "evaluated" VCF to \code{outfile};
otherwise write it to \code{paste0(vcf.name, "_evaluated.vcf")}.}

\item{verbose}{If > 0 print a message when starting the number of slices
generated every \code{verbose} slices.}
}
\value{
Invisibly, a list with the elements
\enumerate{
\item The name of the VCF file created.
\item The in-memory representation of the VCF as a \code{data.table}.
\item The directory with the normal sam slices, if \code{unlink.slice.dir} is \code{FALSE}.
\item The directory with the tumor sam slices, if \code{unlink.slice.dir} is \code{FALSE}.
}
}
\description{
Determine whether sequencing reads in fact support DBSs generated by processing a VCF file.
}
\details{
Creates a new VCF file.
This VCF file has no data rows if there were no DBSs to analyze.
Otherwise, this VCF contains the additional columns
\enumerate{
\item \code{Format} contains the fixed string \code{"WtReads:pos1reads:pos2reads:MutReads"}.
\item \code{NreadSupport} With regard to the two positions of the DBS in
the normal BAM, a string with 4 numbers separated by ":", with the numbers
indicating respectively:
\itemize{
\item the number of reads that are reference sequence at
both positions of the DBS,
\item the number of reads that that have the alternative
allele only at the 1st position of the DBS,
\item the number of reads that
have the alternative allele only at the second position of the DBS, and
\item the number
of reads that have the alternative alleles at both positions of the DBS.
}
\item \code{TreadSupport} Information analogous to that in \code{NreadSupport}, for the
tumor BAM.
\item \code{DBSconclusion} A string that describes whether the DBSs is
believable (\code{"True DBS"}), and if not, a string that describes
why not.
}
}
