% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressSamDirectory.R
\name{RegressSAMDirectory}
\alias{RegressSAMDirectory}
\title{Compare the contents of two directories of .sam files}
\usage{
RegressSAMDirectory(old.dir, new.dir)
}
\arguments{
\item{old.dir}{One directory.}

\item{new.dir}{The other directory.

For use as part of automated tests.}
}
\description{
Compare the contents of two directories of .sam files
}
\keyword{internal}
