% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice2ReadSupport.R
\name{Slice2ReadSupport}
\alias{Slice2ReadSupport}
\title{Calculate the support for a putative DBS from the SAM slice containing the overlapping reads.}
\usage{
Slice2ReadSupport(slice.dir, CHROM, POS, REF, ALT)
}
\arguments{
\item{slice.dir}{The directory containing the SAM slices.}

\item{CHROM}{The chromosome identifier.}

\item{POS}{The first position of the DBS.}

\item{REF}{The reference variant.}

\item{ALT}{The alternate variant.}
}
\description{
Calculate the support for a putative DBS from the SAM slice containing the overlapping reads.
}
