% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifyDBSVcf.R
\name{VerifyDBSVcf}
\alias{VerifyDBSVcf}
\title{Check whether analysis of individual reads supports DBSs in a VCF.}
\usage{
VerifyDBSVcf(vcf, Nbam.name, Tbam.name)
}
\arguments{
\item{vcf}{An in-memory representation of a "variant call file", VCF, as a \code{data.frame}.}

\item{Nbam.name}{The name of the BAM file for the normal sample.}

\item{Tbam.name}{The name of the BAM file for the tumor sample.}
}
\value{
An in-memory VCF based on the input \code{vcf} with the additional columns
\code{ID}, \code{NreadSupport}, \code{TreadSupport}, and \code{DBSconclusion}.
If \code{vcf} is empty, return it without adding columns.

XXXXXX MORE MORE
}
\description{
Check whether analysis of individual reads supports DBSs in a VCF.
}
