% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBSConclusion.R
\name{DBSConclusion}
\alias{DBSConclusion}
\title{Examines every row of of DBS VCF already processed by \code{\link{Slice2ReadSupport}} and decides which DBSs are likely real.}
\usage{
DBSConclusion(vcf, germlineCutOff = 0.2, max.half.support.T.reads = 1)
}
\arguments{
\item{vcf}{An in memory DBS VCF (as a \code{data.frame})
already processed by \code{\link{Slice2ReadSupport}}  (so that the fields
\code{vcf$NreadSupport} and \code{vcf$TreadSupport} are populated).}

\item{germlineCutOff}{If this proportion of normal reads show one or the
other variant (or both variants), consider this a germline variant or
partial germline variant.}

\item{max.half.support.T.reads}{Do not tolerate more than this number of reads in the tumor
that support one but not both mutated positions.}
}
\value{
A VCF with the field \code{DBSconclusion} populated.
}
\description{
Examines every row of of DBS VCF already processed by \code{\link{Slice2ReadSupport}} and decides which DBSs are likely real.
}
\keyword{internal}
