% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG_read_table_and_evaluate_DBS.R
\name{PCAWG_read_table_and_evaluate_DBS}
\alias{PCAWG_read_table_and_evaluate_DBS}
\title{Read a table that specifies how to process VCFs and mini-BAMS to evaluate DBS calls, for PCAWG Collaboratory data.}
\usage{
PCAWG_read_table_and_evaluate_DBS(
  in.table,
  in.vcf.dir,
  minibam.dir,
  out.vcf.dir = in.vcf.dir,
  bam.suffix = "_dbs_srt",
  verbose = 1
)
}
\arguments{
\item{in.table}{the file path of the table to process; in production,
.../DBSverify/data-raw/collaboratory_bams_full_2021_07_13.csv,
for testing .../DBSverify/data-raw/short_collaboratory_bams.csv.}

\item{in.vcf.dir}{The path to the directory containing the DBS VCF files.}

\item{minibam.dir}{The path to the directory containing the mini BAMs.}

\item{out.vcf.dir}{The path to the directory in which to put the "evaluated"
DBS VCF files.}

\item{bam.suffix}{String to add to end of BAM file name; depends on
the conventions used by the script (run on the Collaboratory)
that generated the miniBAMs.}

\item{verbose}{If > 0 generate some progress messages.}
}
\description{
Read a table that specifies how to process VCFs and mini-BAMS to evaluate DBS calls, for PCAWG Collaboratory data.
}
\details{
This is a specialized function for processing
PCAWG data from the ICGC (International Cancer
Genome) "Collaboratory" cloud computing system, once the
miniBAMs have been created in the Collaboratory and
downloaded. The \code{in.table} and associated BED files
were used to specify the contents of the miniBAMs.
The result consists of the "evaluated" DBS VCF files.
The naming of the input and output VCF files and the mini BAMs is governed
by the contents of \code{in.table}, with the VCF file names
incorporating the "aliquot_id" and the miniBAM names
based on the icgc_donor_id and the
T_Specimen ID and N_Specimen ID.
}
