% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_DBS_VCF_and_BAMs_to_verify_DBSs.R
\name{Read_DBS_VCF_and_BAMs_to_verify_DBSs}
\alias{Read_DBS_VCF_and_BAMs_to_verify_DBSs}
\title{Determine whether sequencing reads in fact support (candidate) DBSs present in a VCF file.}
\usage{
Read_DBS_VCF_and_BAMs_to_verify_DBSs(
  input.vcf,
  Nbam.name,
  Tbam.name,
  N.slice.dir = tempfile(),
  T.slice.dir = tempfile(),
  unlink.slice.dir = TRUE,
  exclude.SBSs = TRUE,
  verbose = 0,
  outfile = NULL,
  filter.status = "PASS"
)
}
\arguments{
\item{input.vcf}{If a character string, then the path to a VCF file; otherwise
A a single VCF "file" as a data.frame or similar object.}

\item{Nbam.name}{The name of the BAM file for the normal sample corresponding to \code{vcf.name}.}

\item{Tbam.name}{The name of the BAM file for the tumor sample corresponding to \code{vcf.name}.}

\item{N.slice.dir}{Directory for the slices of the normal BAM.
Created if necessary.}

\item{T.slice.dir}{Directory for the slices of the tumor BAM.
Created if necessary. Must be different than \code{N.slice.dir}.}

\item{unlink.slice.dir}{If \code{TRUE} unlink \code{N.slice.dir}
and \code{T.slice.dir} before return.}

\item{exclude.SBSs}{If \code{TRUE} silently filter out (exclude)
SBSs in the input VCF. This makes sense if the the VCF is from
a caller (like Mutect or the Hartwig Medical Foundation caller)
that calls both SBSs and DBS.}

\item{verbose}{If > 0 print a message when starting the number of slices
generated every \code{verbose} slices.}

\item{outfile}{If not \code{NULL} then write the "evaluated" VCF to \code{outfile};
otherwise write it to \code{paste0(input.vcf(vcf.name, "_evaluated.vcf")}. Must be
non-\code{NULL} if \code{input.vcf} is not a file path.}

\item{filter.status}{If not \code{NULL} only keep rows where the FILTER column
in the VCF is equal to \code{filter.status}.}
}
\value{
Invisibly, a list with the elements
\enumerate{
\item The name of the DBS-only VCF file created.
\item The in-memory representation of the DBS VCF as a \code{data.table}.
\item The name of the directory with the normal SAM slices, if \code{unlink.slice.dir} is \code{FALSE}.
\item The name of the directory with the tumor SAM slices, if \code{unlink.slice.dir} is \code{FALSE}.
}
}
\description{
Determine whether sequencing reads in fact support (candidate) DBSs present in a VCF file.
}
\details{

}
