% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSamfile.R
\name{ReadSamfile}
\alias{ReadSamfile}
\title{Read a "SAM file", discarding some reads that cannot be interpreted for our purposes}
\usage{
ReadSamfile(filename, check.CIGAR = TRUE)
}
\arguments{
\item{filename}{The name of the SAM file to read.}

\item{check.CIGAR}{Include only reads for which the CIGAR string
is only \code{\\d+M} (one or more digits followed M, and nothing else before or after).}
}
\value{
A \code{data.frame} with \code{colnames} for the first 11
columns, with \code{colnames} as specified in
https://en.wikipedia.org/wiki/SAM_(file_format)
and one row per read.
}
\description{
Read a "SAM file", discarding some reads that cannot be interpreted for our purposes
}
\details{
SAM stands for "Sequence Alignment Map", a text file that
represents aligned next-generation sequencing reads
(https://en.wikipedia.org/wiki/SAM_(file_format)).
Only keep reads that statisfy certain conditions:
\itemize{
\item Mate pair maps to same chromosome
\item Mapping quality >= 30
\item "FLAG" < 256
}
}
\keyword{internal}
