% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSamfile.R
\name{ReadSamfile}
\alias{ReadSamfile}
\title{Read a "SAM file", discarding some reads that are uninterpretable for our purposes}
\usage{
ReadSamfile(filename)
}
\arguments{
\item{filename}{The name of the SAM file to read.}
}
\value{
A \code{data.frame} with \code{colnames} for the first 11
columns, with \code{colnames} as specified in
https://en.wikipedia.org/wiki/SAM_(file_format)
and one row per read.
}
\description{
SAM stands for "Sequence Alignment Map", a text file that
represents aligned next-generation sequencing reads
(https://en.wikipedia.org/wiki/SAM_(file_format))
}
