% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSamfile.R
\name{ReadSamfile}
\alias{ReadSamfile}
\title{Read a "SAM file", discarding some reads that cannot be interpreted for our purposes.}
\usage{
ReadSamfile(filename)
}
\arguments{
\item{filename}{The name of the SAM file to read.}
}
\value{
A a list with the elements:
\itemize{
\item \code{good.reads}, \code{data.frame} with
with column names for the
first 11 columns as specified in
https://en.wikipedia.org/wiki/SAM_(file_format)
with one row per read. The result contains only the
reads that meet certain conditions.
\item \code{total.depth}, the initial depth including
"bad" reads
\item \code{reads.with.bad.FLAG}. These are reads
with FLAG >= 256, which marks reads that
(i) are "not primary alignment"
(ii) failed vendor QC
(iii) are PCR or optical duplicates
(iv) are supplementary alignments (e.g. split, split /inverted read).
See https://broadinstitute.github.io/picard/explain-flags.html
\item \code{reads.with.bad.CIGAR}, reads with CIGAR string
that indicates an indel in the read or clipping.
\item \code{reads.with.bad.MAPQ}, reads with MAPQ < 30.
\item \code{reads.with.bad.Mate_CHROM}, reads with a
mate on a different chromosome.
}
}
\description{
Read a "SAM file", discarding some reads that cannot be interpreted for our purposes.
}
\details{
SAM stands for "Sequence Alignment Map", a text file that
represents aligned next-generation sequencing reads
(https://en.wikipedia.org/wiki/SAM_(file_format)).
Only keep reads that satisfy certain conditions:
\itemize{
\item Mate pair maps to same chromosome
\item Mapping quality >= 30
\item "FLAG" < 256 (see info on the return value
element \code{reads.with.bad.FLAG}, below).
\item The CIGAR string
is only \code{\\d+M} (one or more digits followed M, and nothing
else before or after).
This means there are no insertions or deletions in the read
versus the reference and there is no soft clipping. Insertions
or deletions or clipping shift the read within the
SAM file, after which this function cannot keep track of the
DBS location.
}
}
